refactor(es/streaming): separate Streaming folder into WebSockets and Core folders
