+++ modules/evaluation-server/src/Domain/Core/TokenNumber.cs
@@ -1,4 +1,4 @@
-﻿namespace Domain.Streaming;
+﻿namespace Domain.Core;
 
 public static class TokenNumber
 {
