+++ modules/evaluation-server/src/Domain/WebSockets/Connection.cs
@@ -1,6 +1,6 @@
 ﻿using System.Net.WebSockets;
 
-namespace Domain.Streaming;
+namespace Domain.WebSockets;
 
 public class Connection
 {
