+++ modules/evaluation-server/tests/Domain.UnitTests/Core/TokenTests.cs
@@ -1,6 +1,6 @@
-﻿using Domain.Streaming;
+﻿using Domain.Core;
 
-namespace Domain.UnitTests.Streaming;
+namespace Domain.UnitTests.Core;
 
 public class TokenTests
 {
