+++ modules/evaluation-server/tests/Domain.UnitTests/WebSockets/ConnectionTests.cs
@@ -1,9 +1,9 @@
 ﻿using System.Net.WebSockets;
-using Domain.Streaming;
+using Domain.WebSockets;
 using Moq;
-using Version = Domain.Streaming.Version;
+using Version = Domain.WebSockets.Version;
 
-namespace Domain.UnitTests.Streaming;
+namespace Domain.UnitTests.WebSockets;
 
 public class ConnectionTests
 {
