+++ modules/evaluation-server/tests/Domain.UnitTests/WebSockets/ConnectionTypeTests.cs
@@ -1,6 +1,6 @@
-﻿using Domain.Streaming;
+﻿using Domain.WebSockets;
 
-namespace Domain.UnitTests.Streaming;
+namespace Domain.UnitTests.WebSockets;
 
 public class ConnectionTypeTests
 {
